# openapi.yaml
openapi: 3.0.3
info:
  title: OOH Audio Sync API
  description: Backend APIs for the DreamBricks Audio-Video Sync Totem system.
  version: "1.0.0"
servers:
  - url: http://localhost:8001
    description: Local Node.js Server
paths:
...
  /api/media/upload-url:
    post:
      summary: Sign a Direct Upload (S3 mode)
      description: |
        Returns a short-lived (10 min) presigned URL to PUT one file straight to the bucket,
        bound to its content type and size. In local mode returns `{ "direct": false }`:
        upload through `POST /api/media` instead. After the PUT, call `/api/media/upload-complete`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [filename, size]
              properties:
                filename: { type: string, example: "promo.mp4" }
                size: { type: integer, description: "Bytes" }
                overwrite: { type: boolean }
                replace: { type: string, description: "Existing file whose content is replaced (same extension)" }
      responses:
        '200':
          description: "`{ direct, filename, url, headers, expires_in }` — send the PUT with exactly these headers"
        '400':
          description: Invalid name or empty file
        '404':
          description: "`replace` file not found"
        '409':
          description: "File exists (send `overwrite: true`)"
        '413':
          description: Larger than MAX_UPLOAD_MB
  /api/media/upload-complete:
    post:
      summary: Register a Direct Upload (S3 mode)
      description: Checks the file is in the bucket and adds it to the library.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filename: { type: string }
      responses:
        '201':
          description: "`{ success, filename, type, size }`"
        '404':
          description: The file is not in the bucket
  /api/media/split:
    post:
      summary: Upload Video With Audio and Split
      description: |
        Uploads a video that has audio and saves two files in `assets/`:
        `<name>_video.<mp4|webm>` (no audio) and `<name>_audio.mp3` (aligned to the video timeline).
        Videos whose codec browsers can't play (e.g. HEVC/ProRes .mov) are re-encoded to H.264 MP4.
        The original upload is not kept. Requires ffmpeg/ffprobe on the server.
      parameters:
        - name: filename
          in: query
          required: true
          schema:
            type: string
          example: campanha.mov
        - name: overwrite
          in: query
          required: false
          description: Set to `1` to replace existing output files.
          schema:
            type: string
            enum: ["1"]
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '201':
          description: Split done
          content:
            application/json:
              schema:
                type: object
                properties:
                  video: { type: string, example: campanha_video.mp4 }
                  audio: { type: string, example: campanha_audio.mp3 }
                  transcoded: { type: boolean }
                  duration: { type: number, example: 30.0 }
        '400':
          description: Not a video, or the video has no audio track
        '409':
          description: Output files already exist (`files` lists them)
        '413':
          description: File exceeds the size limit
  /api/media/{filename}:
    parameters:
      - name: filename
        in: path
        required: true
        schema:
          type: string
        example: promo.mp4
    put:
      summary: Replace File Content
      description: Overwrites an existing file with the raw request body, keeping its name.
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: File replaced
        '404':
          description: File not found
        '413':
          description: File exceeds the size limit
    patch:
      summary: Rename File
      description: |
        Renames a file (the name is sanitized; a video must stay a video, an audio an audio).
        Totems using the file are updated to the new name, and online totems switch video immediately.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filename: { type: string, example: new-name.mp4 }
      responses:
        '200':
          description: Renamed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  filename: { type: string }
                  updated_totems: { type: array, items: { type: string } }
        '400':
          description: Invalid name or type change
        '404':
          description: File not found
        '409':
          description: Another file already has this name
    delete:
      summary: Delete File
      description: Deletes a file. Refused while any totem is configured to use it.
      responses:
        '200':
          description: Deleted
        '404':
          description: File not found
        '409':
          description: File is in use (`used_by` lists the totems)
  /api/media:
    get:
      summary: List Media Files
      description: All videos and audios in `assets/`, with size, last modified date and the totems using each file.
      responses:
        '200':
          description: Media list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    filename: { type: string, example: promo.mp4 }
                    type: { type: string, enum: [video, audio] }
                    size: { type: integer }
                    modified: { type: string, format: date-time }
                    used_by: { type: array, items: { type: string } }
    post:
      summary: Upload Video or Audio
      description: |
        Saves a video (.mp4, .webm) or audio (.mp3, .wav, .ogg) file into the `assets/` directory.
        The request body is the raw file bytes. The filename is sanitized (accents removed,
        unsupported characters replaced with `_`). Max size is `MAX_UPLOAD_MB` (default 500 MB).
      parameters:
        - name: filename
          in: query
          required: true
          schema:
            type: string
          example: promo.mp4
        - name: overwrite
          in: query
          required: false
          description: Set to `1` to replace an existing file with the same name.
          schema:
            type: string
            enum: ["1"]
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '201':
          description: File saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  filename: { type: string, example: promo.mp4 }
                  type: { type: string, enum: [video, audio] }
                  size: { type: integer, example: 1048576 }
        '400':
          description: Invalid filename/extension, empty file or aborted upload
        '409':
          description: A file with this name already exists (retry with `overwrite=1`)
        '413':
          description: File exceeds the size limit
...
  /media/{filename}:
    get:
      summary: Stream Media File
      description: Streams a video or audio file with support for HTTP Range requests.
      parameters:
        - in: path
          name: filename
          required: true
          schema:
            type: string
          description: The filename of the media to stream
      responses:
        '200':
          description: Full file content
          content:
            video/mp4:
              schema:
                type: string
                format: binary
        '206':
          description: Partial content (Range response)
          content:
            video/mp4:
              schema:
                type: string
                format: binary
        '400':
          description: Invalid filename path
        '404':
          description: File not found
        '416':
          description: Range not satisfiable
...